django_app/requirements.txt
Django>=3.2,<4.0
djangorestframework
mysqlclient
django-cors-headers
django-filter
djangorestframework-simplejwt
